x86_64: map and unmap pages

+++ src/include/platform/platform.h
@@ -23,5 +23,5 @@
 int platformCPUSetup();         // very early setup for one CPU
 int platformPagingSetup();      // paging setup for virtual memory management
 uintptr_t platformGetPage(int *, uintptr_t);     // get physical address and flags of a page
-int platformMapPage(uintptr_t, uintptr_t, int);         // map a physical address to a virtual address
+uintptr_t platformMapPage(uintptr_t, uintptr_t, int);    // map a physical address to a virtual address
 int platformUnmapPage(uintptr_t);               // and vice versa
